requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
gunicorn>=21.2.0

